+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Changes
+* Prep internal testsuites for having multiple internal testsuites, one per language
 
 # 1.18.8
 ### Changes