+++ core/docs/changelog.md
@@ -1,7 +1,17 @@
 # TBD
 ### Changes
+<<<<<<< HEAD
 * Prep internal testsuites for having multiple internal testsuites, one per 
 * Split the `WaitForEndpointAvailability` api container function to `WaitForEndpointAvailabilityHttpGet` and `WaitForEndpointAvailabilityHttpPost`
+=======
+* Prep internal testsuites for having multiple internal testsuites, one per language
+* Switch to using `container-engine-lib` for `DockerManager`
+* Switch to using `kurtosis-core-launcher-lib`
+
+### Removals
+* Removed the API container `docker_api` package
+* Removed the `ApiContainerLauncher` class here, in favor of the one from `kurtosis-core-launcher-lib`
+>>>>>>> develop
 
 # 1.18.8
 ### Changes